pyproject.toml
[project]
name = "pulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "kafka-python>=2.3.0",
]
